This makes UnswizzleTexture two times faster. It is the main bottleneck in video decoding.

| What | Before | After |
|-|-|-|
| Pokemon intro movie | 27 FPS | 30 FPS |
| Kirby intro movies | 29 FPS | 44 FPS |
| Kirby first world | 25 FPS | 30 FPS |
| Kirby first level | 22 FPS | 27 FPS |
| Link's Awakening intro movie | 12 FPS | 21 FPS |
| Animal crossing set-up | 25 FPS | 30 FPS |
| Animal crossing intro movie | 29 FPS | 30 FPS |
| Animal crossing in-game | 18 FPS | 23 FPS |