@ReinUsesLisp 2021-08-01
Does the performance difference stay if you just call std::memcpy but leave BYTES_PER_PIXEL as a constant? That way you can avoid the UB. Consider using assume_aligned if that doesn't work.