@yzct12345 2021-08-01
> std::memcpy but leave BYTES_PER_PIXEL

Oh, so that's what @lioncash meant.

> memcpy can be optimized out

Wow!

> Does the performance difference

The 2x improvement is still there. Specifically I measured an insignificant improvement in my unit test benchmark but an insignificant worsening of Link's Awakening.

> alignment requirements

I tested with `if (((size_t) dst | (size_t) src) & 3) abort()` and it didn't crash. This means the actual buffers we use are aligned. Additionally on x86 but not ARM unaligned access doesn't crash but is much slower and I would have noticed that.

> assume_aligned

I tested the first suggestion and it works fine. The disassembly looks about the same as my version.