@gidoly 2021-08-01
world ends with you video test
current early access 25 fps ish with laggyness
![yuzu_dX9T7pe9F5](https://user-images.githubusercontent.com/66776795/127760851-6df9114c-9057-47fa-b622-e55575d7cf17.jpg)
with this pr
![video_+_astc_EUOFR26mnX](https://user-images.githubusercontent.com/66776795/127760893-a66ff59c-ea72-4d56-9b89-5afc3d7359ee.jpg)