@yzct12345 2021-08-01
> locally merged to make ZLA

So that's where the flickering was coming from! Thank you! I thought I merged all of early access but I must have missed that one while rebasing.

> please confirm me if using your GitHub name is fine

Yes, I confirm it's fine.