@bylaws 2021-08-01
It's worth checking if the LUT for the Morton swizzle is actually beneficial, at least for us on ARM64 using a LUT decreased the speed heavily as bitfield ops were much cheaper. No idea if this will be the case on x86 though