@yzct12345 2021-08-01
I tried removing the LUT and optimizing out the multiplications but performance was worse. My unit test lost 20% performance from 4 seconds to 5 seconds.