@yzct12345 2021-08-01
It's not complicated. I simply call it over and over.

```
diff --git a/src/video_core/CMakeLists.txt b/src/video_core/CMakeLists.txt
index 1eb67c051..82cbfdcf8 100644
--- a/src/video_core/CMakeLists.txt
+++ b/src/video_core/CMakeLists.txt
@@ -229,6 +229,10 @@ endif()
 target_include_directories(video_core PRIVATE ${FFmpeg_INCLUDE_DIR})
 target_link_libraries(video_core PRIVATE ${FFmpeg_LIBRARIES})
 
+add_executable(swizzlebench textures/swizzlebench.cpp)
+
+target_link_libraries(swizzlebench PUBLIC video_core)
+
 add_dependencies(video_core host_shaders)
 target_include_directories(video_core PRIVATE ${HOST_SHADERS_INCLUDE})
 target_include_directories(video_core PRIVATE sirit ../../externals/Vulkan-Headers/include)
diff --git a/src/video_core/textures/swizzlebench.cpp b/src/video_core/textures/swizzlebench.cpp
new file mode 100644
index 000000000..d2220b76d
--- /dev/null
+++ b/src/video_core/textures/swizzlebench.cpp
@@ -0,0 +1,29 @@
+#include <iostream>
+#include "video_core/textures/decoders.h"
+
+/*
+static u32 st(u32 x, u32 y) {
+    return ((x & 32) << 3) | ((x & 16) << 1) | (x & 15) | ((y & 6) << 5) | ((y & 1) << 4);
+}
+
+using Tegra::Texture::SWIZZLE_TABLE;
+*/
+int main() {
+    std::vector<u8> input(8294400);
+    std::vector<u8> output(8355840);
+    for (int i = 1000; --i;) {
+        Tegra::Texture::UnswizzleTexture(output, input, 4, 1920, 1080, 1, 1, 0, 4);
+    }
+    //    for (u32 y = 0; y < SWIZZLE_TABLE.size(); y++) {
+    //        auto table = SWIZZLE_TABLE[y];
+    //        for (u32 x = 0; x < table.size(); x++) {
+    //            auto expect = table[x];
+    //            auto got = st(x, y);
+    //            if (expect != got) {
+    //                LOG_CRITICAL(Debug, "Expect {} got {}", expect, got);
+    //                abort();
+    //            }
+    //        }
+    //    }
+    return 0;
+}
```

Downclock CPU to 2 Ghz and run `time build/bin/swizzlebench`.