@Wunkolo 2021-08-02
Just to add a data-point of something I just tried: I replaced the Morton-LUT with a [BMI2-based](https://en.wikipedia.org/wiki/X86_Bit_manipulation_instruction_set#BMI2_(Bit_Manipulation_Instruction_Set_2)) implementation using only two [pdep](https://www.felixcloutier.com/x86/pdep)s and an `or`.

The pdep implementation is equivalent to `SWIZZLE_TABLE`.

```cpp
inline u32 TegraMorton2D(u8 X, u8 Y)
{
        const u16 SpreadX = _pdep_u32(X, 0b100101111);
        const u16 SpreadY = _pdep_u32(Y, 0b011010000);
        return SpreadX | SpreadY;
}
```

Using your `swizzlebench` above this is what I got on mainline
```
hyperfine ./bin/swizzlebench
Benchmark #1: ./bin/swizzlebench
  Time (mean ± σ):      7.365 s ±  0.014 s    [User: 7.412 s, System: 0.026 s]
  Range (min … max):    7.351 s …  7.393 s    10 runs
```

this is what I got after checking out this PR:
```
hyperfine ./bin/swizzlebench
Benchmark #1: ./bin/swizzlebench
  Time (mean ± σ):      2.684 s ±  0.011 s    [User: 2.729 s, System: 0.025 s]
  Range (min … max):    2.661 s …  2.696 s    10 runs
```

This is what I got after checking out this PR and replacing all uses of `SWIZZLE_TABLE` with my `BMI2`-based implementation:
```
hyperfine ./bin/swizzlebench
Benchmark #1: ./bin/swizzlebench
  Time (mean ± σ):      2.679 s ±  0.013 s    [User: 2.723 s, System: 0.029 s]
  Range (min … max):    2.653 s …  2.694 s    10 runs
```

It's just a bit faster.

Another thing to possibly check out is if an [AVX2-based scatter-gather](https://software.intel.com/sites/landingpage/IntrinsicsGuide/#text=gather&techs=AVX2) would speed up the inner `column` loop by doing all the LUT-lookups, offset calculations, and memcpys in parallel simd-sized batches.

Depends how the codebase feels about having arch-specific intrinsics.